Update readme
The sample CALAND() run for Alternative A scenario had an error: the argument for non-regeneration (NR_Dist) was set to 120, which is not how they were run for the Implementation Plan and will generate filenames that don't correspond to the subsequent sample plot_caland() and plot_scen_types() runs. NR_Dist should be set to -1 (maximum regeneration) for Alternative A, which emulates maximum reforestation potential relative to a baseline of maximum non-regeneration.