Update readme
Editing a couple things in the CALAND references. There was only one citation for CALAND v3, which was specifically for the technical documentation. I think we should also have one for citing the code too, both of which have the same DOI. I separated them into 2 citations here, but I could also see doing it where we combine both and call it "technical documentation and code", and cite the DOI.